backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
kombu==5.5.3
numpy==2.2.6
opencv-python==4.11.0.86
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
redis==6.1.0
ruff==0.11.10
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
